resources/legacy-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weekly-report"
version = "0.1.0"
description = "自动生成周/月报的脚本"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-docx>=0.8.11",
]

[tool.setuptools]
py-modules = ["weekly"]
